requirements.txt
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools~=5.3.3
certifi~=2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain~=0.3.25
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith~=0.3.45
MarkupSafe==3.0.2
narwhals==2.0.1
numpy~=1.26.4
openai==1.98.0
orjson~=3.10.14
ormsgpack==1.10.0
packaging~=23.2
pandas~=2.2.2
pillow~=10.4.0
protobuf~=6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv~=1.0.0
pytz==2025.2
PyYAML~=6.0.1
referencing==0.36.2
regex==2025.7.31
requests~=2.32.5
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio~=1.3.0
SQLAlchemy==2.0.42
sqlite-vec==0.1.6
streamlit~=1.49.1
tenacity~=9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado~=6.5
tqdm~=4.66.4
typing-inspection==0.4.1
typing_extensions~=4.15.0
tzdata==2025.2
urllib3~=2.5.0
watchdog==6.0.0
xxhash==3.5.0
zstandard==0.23.0
jupyter~=1.0.0
pip~=24.0
panel~=1.4.4
zlib~=1.3.1
pysocks~=1.7.1
openssl~=3.5.2
pyopenssl~=24.0.0
cryptography~=44.0.1
docutils~=0.18.1
sphinx~=7.3.7
filelock~=3.13.1
rsa~=4.7.2
pyasn1~=0.4.8
primp~=0.15.0
lxml~=6.0.1
grpcio~=1.71.0
beautifulsoup4~=4.12.3
bleach~=4.1.0
cssselect~=1.2.0
dask~=2024.5.0
rich~=13.9.4
keras~=3.11.0
ipywidgets~=7.8.1
ipython~=8.25.0
notebook~=7.2.2
yaml~=0.2.5
pytest~=7.4.4
pygments~=2.15.1
brotli~=1.0.9
proto-plus~=1.26.1
aiohttp~=3.12.14
pyasn1-modules~=0.2.8
cffi~=1.16.0
google-api-python-client~=2.181.0
httplib2~=0.31.0
filetype~=1.2.0
pyparsing~=3.0.9
mypy~=1.17.1
email_validator~=2.3.0
email-validator~=2.3.0
deprecated~=1.2.13
chardet~=4.0.0
psutil~=5.9.0
uritemplate~=4.2.0
transformers~=4.53.0
defusedxml~=0.7.1
google-auth-httplib2~=0.2.0
duckduckgo-search~=8.1.1
python-multipart~=0.0.20
langchain-google-genai~=2.0.10
langchain-community~=0.3.27